Organización

+++ Basicos/costevidacaprichos.py
@@ -1,6 +1,6 @@
 print ("Hola, vamos a proceder a calcular tu tiempo de vida que te costará si te compras el capricho")
 coste_aparato=float(input("introduce el coste del aparato : "))
-salario=float(2000)
+salario=float(1000)
 precioporhora=float((salario/21)/7.5)
 horas=float(coste_aparato/precioporhora)
 print("El precio por hora es :",precioporhora)
